feat: use phantomjs npm module
When the phantomjs npm module is installed it includes the
phantomjs binary native to the OS. Using this binary allows
Karma to launch tests via PhantomJS "out of the box," without
the need for users to install PhantomJS manually or worry
about configuring the path to their PhantomJS executable.

Users are still free to install their own copy of PhantomJS
and override the PHANTOMJS_BIN environment variable.

Closes #1